requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.1
distlib==0.3.1
filelock==3.0.12
future==0.18.2
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
lxml==4.6.1
multidict==4.7.6
numpy==1.19.4
Pillow==8.0.1
python-dotenv==0.15.0
requests==2.24.0
six==1.15.0
StackAPI==0.1.12
torch==1.5.1+cpu
torchvision==0.6.1+cpu
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.1.0
virtualenv-clone==0.5.4
yarl==1.5.1
zipp==3.4.0
